pyproject.toml
[project]
name = "supply-chain-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-api-python-client==2.149.0",
    "langchain==0.3.0",
    "langchain-anthropic==0.2.0",
    "langchain-community==0.3.0",
    "langchain-core==0.3.0",
    "langchain-google-genai==2.0.0",
    "langchain-openai==0.2.0",
    "langsmith==0.1.147",
    "python-dotenv==1.0.1",
]
